Rewrite
The original Github code was written in June 2011, with some updates out
to 2015.

This code has been worked on, on and off since June 2011.  This is a
rewrite of it in Kotlin.  Some problems have gone away, like dex file
65536 method limits.

I have more features already written that I will push up.